Added eigenvalue simulations as a constraint
Initial work on reading a lua input file that contains `Simulations`
and `Constraints` tables.  These tables should be looped over and processed.
Much of the logic needs to be embedded in the table itself.  The exact manner
of how things are handled will likely change when we move to MPI/Coroutine
based solutions